// frontend_angular/src/app/api/model/device.ts
/**
 * Adherent
 * Adherent api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Device { 
    connectionType: Device.ConnectionTypeEnum;
    ipAddress?: string;
    ipv6Address?: string;
    mac: string;
    username?: string;
}
export namespace Device {
    export type ConnectionTypeEnum = 'wired' | 'wireless';
    export const ConnectionTypeEnum = {
        Wired: 'wired' as ConnectionTypeEnum,
        Wireless: 'wireless' as ConnectionTypeEnum
    };
}